+++ magellan-sample-advanced/src/main/java/com/wealthfront/magellan/sample/advanced/tide/TideLocationsScreen.java
@@ -3,10 +3,10 @@ package com.wealthfront.magellan.sample.advanced.tide;
 import android.content.Context;
 import androidx.annotation.DrawableRes;
 
-import com.wealthfront.magellan.Screen;
+import com.wealthfront.magellan.LegacyScreen;
 import com.wealthfront.magellan.sample.advanced.R;
 
-public class TideLocationsScreen extends Screen<TideLocationsView> {
+public class TideLocationsScreen extends LegacyScreen<TideLocationsView> {
 
   public enum TideLocations {
     SAN_FRANCISCO(9414290, R.drawable.san_francisco, "San Francisco"),
